FLL-25 edit code base on feedback
