+++ package.json
@@ -36,6 +36,7 @@
     "eslint-config-next": "15.1.0",
     "json-server": "^1.0.0-beta.3",
     "postcss": "^8",
+    "react-hot-toast": "^2.4.1",
     "tailwindcss": "^3.4.1",
     "typescript": "^5"
   }
